Разбивает утилитарный модуль на несколько модулей: общий, модуль для рендеринга и утилиты для задач
